Extract allocators.h
Extracts allocators.c functions from hodlr.h into allocators.h
