+++ tests/src/test_real_data.c
@@ -19,6 +19,7 @@
 
 #include "../../include/hmat_lib/hodlr.h"
 #include "../../include/hmat_lib/error.h"
+#include "../../include/hmat_lib/allocators.h"
 #include "../../include/internal/blas_wrapper.h"
 
 #include "../include/io.h"
